z3helper.py removed
